Added Django 1.5 and (incomplete) Python3 support.
I updated the code to make it work with python 2.5-3.3 and django 1.3-1.5. For this reason I didn't use django 1.5's helpers but pylib's, that are a bit more verbose but work well.

There are still a bunch of tests that fail with Python 3.x, basically any test that uses fixtures. I won't have time this week to keep working on this, but note that I'm successfully using pytest fixtures with Django and Python3 in django-treebeard, so the problem must be something I couldn't trace in pytest_django.